OSIS-6249 : Déplacement du code
