+++ base/api/serializers/student.py
@@ -27,26 +27,6 @@ from rest_framework import serializers
 
 from base.api.serializers.person import PersonDetailSerializer
 from base.models.student import Student
-from django.urls import reverse
-
-
-class StudentListSerializer(serializers.Serializer):
-    registration_id = serializers.CharField()
-    name = serializers.SerializerMethodField()
-    gender = serializers.SerializerMethodField()
-    select_url = serializers.SerializerMethodField()
-
-    def get_name(self, obj):
-        return str(obj.person)
-
-    def get_gender(self, obj):
-        return obj.person.gender
-
-    def get_select_url(self, obj):
-        return reverse(
-            "student_read",
-            kwargs={'student_id': obj.id}
-        )
 
 
 class StudentSerializer(serializers.ModelSerializer):
