+++ base/urls.py
@@ -43,11 +43,12 @@ import base.views.learning_units.search.proposal
 import base.views.learning_units.search.service_course
 import base.views.learning_units.search.simple
 import base.views.learning_units.update
+import base.views.student.detail
 from attribution.views import attribution
 from base.views import geocoding
 from base.views import learning_achievement, search, user_list
 from base.views import learning_unit, common, institution, organization, academic_calendar, \
-    my_osis, student
+    my_osis
 from base.views import teaching_material
 from base.views.autocomplete import OrganizationAutocomplete, CountryAutocomplete, CampusAutocomplete, \
     EntityAutocomplete, AllocationEntityAutocomplete, AdditionnalEntity1Autocomplete, AdditionnalEntity2Autocomplete, \
@@ -60,7 +61,8 @@ from base.views.learning_units.pedagogy.update import learning_unit_pedagogy_edi
     learning_unit_pedagogy_force_majeure_edit
 from base.views.learning_units.proposal import create, update
 from base.views.learning_units.update import update_learning_unit, learning_unit_edition_end_date
-from base.views.student import StudentSearch
+from base.views.student.list import StudentSearch
+from base.views.student.detail import StudentRead
 from education_group import urls as education_group_urls
 from learning_unit import urls as learning_unit_urls
 
@@ -302,8 +304,8 @@ urlpatterns = [
     url(r'^students/', include([
         path('', StudentSearch.as_view(), name='students'),
         url(r'^(?P<student_id>[0-9]+)/', include([
-            url(r'^$', student.StudentRead.as_view(), name='student_read'),
-            url(r'^picture$', student.student_picture, name='student_picture'),
+            url(r'^$', StudentRead.as_view(), name='student_read'),
+            url(r'^picture$', base.views.student.detail.student_picture, name='student_picture'),
         ]))
     ])),
     url(r'^ajax_select/', include(ajax_select_urls)),
